+++ SDWebImage/UIView+WebCacheOperation.h
@@ -17,20 +17,20 @@
  *  @param operation the operation
  *  @param key       key for storing the operation
  */
-- (void)sd_setImageLoadOperation:(id)operation forKey:(NSString *)key;
+- (void)sd_setImageLoadOperation:(nullable id)operation forKey:(nullable NSString *)key;
 
 /**
  *  Cancel all operations for the current UIView and key
  *
  *  @param key key for identifying the operations
  */
-- (void)sd_cancelImageLoadOperationWithKey:(NSString *)key;
+- (void)sd_cancelImageLoadOperationWithKey:(nullable NSString *)key;
 
 /**
  *  Just remove the operations corresponding to the current UIView and key without cancelling them
  *
  *  @param key key for identifying the operations
  */
-- (void)sd_removeImageLoadOperationWithKey:(NSString *)key;
+- (void)sd_removeImageLoadOperationWithKey:(nullable NSString *)key;
 
 @end